Function polkadot_node_core_pvf_prepare_worker::worker_entrypoint

pub fn worker_entrypoint(
    socket_path: PathBuf,
    worker_dir_path: PathBuf,
    node_version: Option<&str>,
    worker_version: Option<&str>,
)

The entrypoint that the spawned prepare worker should start with.

Parameters

socket_path: specifies the path to the socket used to communicate with the host.

worker_dir_path: specifies the path to the worker-specific temporary directory.

node_version: if Some, is checked against the worker_version. A mismatch results in immediate worker termination. None is used for tests and in other situations when version check is not necessary.

worker_version: see above

Flow

This runs the following in a loop:

Get the code and parameters for preparation from the host.

Start a new child process

Start the memory tracker and the actual preparation in two separate threads.

Wait on the two threads created in step 3.

Stop the memory tracker and get the stats.

Pipe the result back to the parent process and exit from child process.

If compilation succeeded, write the compiled artifact into a temporary file.

Send the result of preparation back to the host, including the checksum of the artifact. If any error occurred in the above steps, we send that in the PrepareWorkerResult.